refactor: apis 통합
