+++ client/src/apis/users.ts
@@ -1,6 +1,18 @@
 import client from '@/apis/utils/client';
 import User, { Profile } from '@/types/User';
 
+interface SignupRequest {
+  myTilAddress: string;
+  nickname: string;
+  email: string;
+  password: string;
+  notification?: User['notification'];
+}
+
+export async function signup(data: SignupRequest): Promise<void> {
+  return client.post('/signup', data).then((res) => res.data);
+}
+
 export interface LoginRequest {
   email: User['email'];
   password: string;
