+++ client/src/components/common/NotificationInput.tsx
